requirements.txt
python-dotenv
Flask
requests
gunicorn 
os 